Organized screens into seperate files
Moved BudgetListScreen and budgetScreen out of App.js and into seperate files. App.js not imports the screen to use it.
