+++ Budgie/screens/budgetScreen.js
@@ -1,0 +1,41 @@
+import React, {Component} from 'react';
+import {
+    Text,
+  } from 'react-native';
+
+export const BudgetScreen = ({ navigation, route }) => {
+    // const category = ({ title, allocated }) => {
+    //   return(
+        
+    //   );
+    // };
+  
+    // const spending = ({ title }) => {}
+    //   return( 
+    //     <View style={styles.item}>
+    //       <Text style={styles.title}>{title}</Text>
+    //     </View>
+    //     );
+    //   };
+  
+  
+    return(
+    //   <SafeAreaView>
+    //     <StatusBar barStyle={'dark-content'} />
+    //     <ScrollView
+    //       contentInsetAdjustmentBehavior="automatic"
+    //       <View>
+    //       <SectionList
+    //         sections={DATA}
+    //         keyExtractor={(item, index) => item + index}
+    //         renderItem={({ item }) => <Item title={item} />}
+    //         renderSectionHeader={({ section: { title } }) => (
+    //       <Text style={styles.header}>{title}</Text>
+    //     )}
+    //   />
+    //     </View>
+    //   </ScrollView>
+    // </SafeAreaView>
+        <Text> Hi </Text>
+    );
+  };
